Working version. Communication with google cloud vision api fixed.

+++ src/main/java/com/alibou/security/project/baza/model/ImageReceipt.java
@@ -15,7 +15,7 @@ public class ImageReceipt {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    Long imageReceiptId;
+    Integer imageReceiptId;
 
     String name;
